QP shim main (#13128)
* git mv cmd/main/*.go pkg/queue/qmain/

Fixes: https://github.com/knative/serving/issues/13127

This first step is sent as a seperate commit to get git to recognize that the files were renamed and not changed.
Changes will be a seperate commit

* iAdd shim main to queue + correct qmain package names

Fixes: https://github.com/knative/serving/issues/13127

* Renamed to pkg/queue/qmain to pkg/queue/shardmain (#13127)

* Change package name following move from pkg/queue/qmain to pkg/queue/sharedmain (#13127)
